First steps for using global settings
The begining for using global settings and a more advanced dictionary
structure for holding results. In the end we should be able to have a
dictionary where we can lookup by cut-ratio and the specific cut.

We want to have this structure:
    {cutRatio : {cutNo : {blobName : color, connectedComponent} }}

Example (two cuts):
    {
     0.618 : {0 : {}, 1 : { a blob}, 2 : {}, 3 : {} },
     0.667 : {0 : { a blob}, 1 : {}, 2 : {}, 3 : {} }
    }

Before we just threw all of the components is the same pile. This way we
can separate the induvidual cuts (which is good).