Refactoring of code regarding the naive method
The naive method have been moved out of the paintingAnalyzer into it's
own file. This is for keeping the file simple and group the method for
the naive method together. It will be a simple task to create a new
method file, importing that into the paintingAnalyzer and then use it.
Finally, the tests have been cleaned up, taking advantage of the
Settings class instead of using some ugly hack.